Added an anti-stalemate feature.
If one player is ahead of the other player for a long time, this change
causes that player to gain an advantage in the form of the losing
player losing a ship. This should prevent cases where it's clear one
player is better, but the penalty for the winning player prevents the
game from ending.