Added MP3 backups for Ogg files.
These are for the Windows builds; for some reason, PyInstaller builds
on Windows can't load Ogg Vorbis files. The good news is that it
errors out and so when that happens we can attempt to load MP3s
instead. (At least in theory; if this doesn't work when I try it on
the Windows computer then I'll revert.)